Move server startup script logic to the main startup script
